requirements.txt
git+https://github.com/minturtle/redis-subscriber.git@v1.0.1
python-dotenv
arxiv
langchain-openai
pymongo